microservice/requirements.txt
fastapi
uvicorn
bcrypt
pyjwt